requirements.txt
pillow>=10.0.0
numpy>=1.24.0
piexif>=1.1.3
click>=8.1.3
requests>=2.31.0
pillow-avif-plugin>=1.4.5
openai>=1.40.0

